requirements.txt
neo4j 
pandas 
sentence-transformers 
langchain 
langchain-openai 
langchain-neo4j
python-dotenv